+++ client/constants/actions.tsx
@@ -2,11 +2,9 @@
 export const LOGIN_FAILED = 'LOGIN_FAILED'
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESSFULLY'
 
+
 export const VERIFY_GOOGLE_TOKEN_SUCCESS = 'VERIFY_GOOGLE_TOKEN_SUCCESS'
 export const VERIFY_GOOGLE_TOKEN_FAILED = 'VERIFY_GOOGLE_TOKEN_FAILED'
-
-export const VERIFY_TWITTER_TOKEN_SUCCESS = 'VERIFY_TWITTER_TOKEN_SUCCESS'
-export const VERIFY_TWITTER_TOKEN_FAILED = 'VERIFY_TWITTER_TOKEN_FAILED'
 
 
 export const REGISTER_FAILED = 'REGISTER_FAILED'
@@ -20,3 +18,10 @@ export const LOGOUT = 'LOGOUT'
 // Spotify
 
 // Twitter
+export const VERIFY_TWITTER_TOKEN_SUCCESS = 'VERIFY_TWITTER_TOKEN_SUCCESS'
+export const VERIFY_TWITTER_TOKEN_FAILED = 'VERIFY_TWITTER_TOKEN_FAILED'
+export const TWITTER_LOGIN_FAILED = 'TWITTER_LOGIN_FAILED'
+export const TWITTER_LOGIN_SUCCESS = 'TWITTER_LOGIN_SUCCESSFULLY'
+export const GET_TWEETS_SUCCESS = 'TWEETS_SUCCESS'
+export const GET_TWEETS_FAILED = 'TWEETS_FAILED'
+
