+++ client/store/reducers/index.tsx
@@ -1,5 +1,6 @@
 import { combineReducers } from 'redux'
 import userReducer from './userReducer'
+import TwitterReducer from "./twitterReducer";
 
 export interface State {
   loading: boolean
@@ -8,4 +9,5 @@ export interface State {
 
 export default combineReducers({
   user: userReducer,
+  twitterUser: TwitterReducer
 })
